server/pyproject.toml
[tool.poetry]
name = "callability-server"
version = "0.4.0"
description = ""
authors = ["Amos Joshua"]
readme = "README.md"
packages = [{include = "callability_server"}]
package-mode = true

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.67.0"
grpcio_tools = "^1.67.0"
grpcio_reflection = "^1.67.0"
firebase-admin = "^6.5.0"
protobuf = "^5.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7"
black = "^23.1.0"
pre-commit = "^3.1.0"
flake8 = "^6.0.0"
mypy = "^1.0.1"
isort = "^5.12.0"

[tool.poetry.scripts]
grpc-server = 'callability_server:main.cli'
check-health = 'callability_server:check_health.cli'

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.black]
line-length = 120

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.isort]
profile = "black"
src_paths = ["callability_server", "tests"]
